Add a template for AIAA formatted papers to the list.
I'm not completely sure where the most appropriate location for this template is. It's not an "official quality" template,
but handles the vast majority of the tricky legwork for getting AIAA formatting working. It took quite a while to write, so I'd
like to save everyone else's time in the future.